restructuon speech list component ;
